create DeliveryAddresses folder and move related files
